backend/requirements.txt
# Intelli-Credit-Aveva Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
langgraph>=0.2.0
langsmith>=0.1.0
qdrant-client>=1.7.0
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
xgboost>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
fpdf2>=2.7.0
